event-driven-demo/order-service/requirements.txt
flask==2.3.3
kafka-python==2.0.2
